extract language registry
