+++ domain/language.ts
@@ -1,16 +1,23 @@
-import type { Language } from '@/domain/settings';
-import { translations } from './index';
+export const SUPPORTED_LANGUAGES = {
+  de: true,
+  en: true,
+  hi: true,
+  pl: true,
+  'zh-CN': true,
+} as const;
+
+export type Language = keyof typeof SUPPORTED_LANGUAGES;
 
 const DEFAULT_LANGUAGE: Language = 'en';
 
 export function selectLanguage(browserLanguages: readonly string[]): Language {
   for (const browserLanguage of browserLanguages) {
-    if (browserLanguage in translations) {
+    if (browserLanguage in SUPPORTED_LANGUAGES) {
       return browserLanguage as Language;
     }
 
     const baseLanguage = browserLanguage.split('-')[0];
-    if (baseLanguage in translations) {
+    if (baseLanguage in SUPPORTED_LANGUAGES) {
       return baseLanguage as Language;
     }
     if (baseLanguage === 'zh') {
@@ -22,5 +29,5 @@ export function selectLanguage(browserLanguages: readonly string[]): Language {
 }
 
 export function getSupportedLanguage(language: unknown): Language | undefined {
-  return typeof language === 'string' && language in translations ? (language as Language) : undefined;
+  return typeof language === 'string' && language in SUPPORTED_LANGUAGES ? (language as Language) : undefined;
 }
