+++ infrastructure/browser/language.ts
@@ -1,5 +1,4 @@
-import type { Language } from '@/domain/settings';
-import { selectLanguage } from '@/shared/i18n/language';
+import { type Language, selectLanguage } from '@/domain/language';
 
 export function detectBrowserLanguage(): Language {
   const browserLanguages = typeof navigator !== 'undefined' ? navigator.languages : [];
